requirements.txt
streamlit>=1.28.0
boto3>=1.34.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
statsmodels>=0.14.0
plotly>=5.17.0
python-dotenv>=1.0.0
streamlit-authenticator>=0.2.3
bcrypt>=4.1.0
pyyaml>=6.0.0
openai>=1.6.0
openpyxl>=3.1.0